api/requirements.txt
azure-core==1.30.2
azure-cosmos==4.7.0
azure-functions==1.20.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
typing-extensions==4.12.2
urllib3==2.2.2
